+++ src/Delegate/NotFoundDelegate.php
@@ -7,106 +7,12 @@
 
 namespace Zend\Expressive\Delegate;
 
-use Fig\Http\Message\StatusCodeInterface;
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Zend\Expressive\Template\TemplateRendererInterface;
+use Zend\Expressive\Handler\NotFoundHandler;
 
-class NotFoundDelegate implements DelegateInterface
+/**
+ * @deprecated since 2.2.0; to be removed in 3.0.0. Use
+ *     Zend\Expressive\Handler\NotFoundHandler instead, available since 2.2.0.
+ */
+class NotFoundDelegate extends NotFoundHandler
 {
-    const TEMPLATE_DEFAULT = 'error::404';
-    const LAYOUT_DEFAULT = 'layout::default';
-
-    /**
-     * @var TemplateRendererInterface
-     */
-    private $renderer;
-
-    /**
-     * This duplicates the property in StratigilityNotFoundHandler, but is done
-     * to ensure that we have access to the value in the methods we override.
-     *
-     * @var ResponseInterface
-     */
-    protected $responsePrototype;
-
-    /**
-     * @var string
-     */
-    private $template;
-
-    /**
-     * @var string
-     */
-    private $layout;
-
-    /**
-     * @param ResponseInterface $responsePrototype
-     * @param TemplateRendererInterface $renderer
-     * @param string $template
-     * @param string $layout
-     */
-    public function __construct(
-        ResponseInterface $responsePrototype,
-        TemplateRendererInterface $renderer = null,
-        $template = self::TEMPLATE_DEFAULT,
-        $layout = self::LAYOUT_DEFAULT
-    ) {
-        $this->responsePrototype = $responsePrototype;
-        $this->renderer = $renderer;
-        $this->template = $template;
-        $this->layout = $layout;
-    }
-
-    /**
-     * Creates and returns a 404 response.
-     *
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    public function process(ServerRequestInterface $request)
-    {
-        if (! $this->renderer) {
-            return $this->generatePlainTextResponse($request);
-        }
-
-        return $this->generateTemplatedResponse($request);
-    }
-
-    /**
-     * Generates a plain text response indicating the request method and URI.
-     *
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    private function generatePlainTextResponse(ServerRequestInterface $request)
-    {
-        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
-        $response->getBody()
-            ->write(sprintf(
-                'Cannot %s %s',
-                $request->getMethod(),
-                (string) $request->getUri()
-            ));
-        return $response;
-    }
-
-    /**
-     * Generates a response using a template.
-     *
-     * Template will receive the current request via the "request" variable.
-     *
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    private function generateTemplatedResponse(ServerRequestInterface $request)
-    {
-        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
-        $response->getBody()->write(
-            $this->renderer->render($this->template, ['request' => $request, 'layout' => $this->layout])
-        );
-
-        return $response;
-    }
 }