+++ src/Middleware/ImplicitHeadMiddleware.php
@@ -1,55 +1,30 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
  */
 
 namespace Zend\Expressive\Middleware;
 
-use Fig\Http\Message\RequestMethodInterface as RequestMethod;
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface as ServerMiddlewareInterface;
 use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\Stream;
-use Zend\Expressive\Router\RouteResult;
+use Zend\Expressive\Router\Middleware\ImplicitHeadMiddleware as BaseImplicitHeadMiddleware;
 
 /**
  * Handle implicit HEAD requests.
  *
- * Place this middleware after the routing middleware so that it can handle
- * implicit HEAD requests -- requests where HEAD is used, but the route does
- * not explicitly handle that request method.
+ * This is an extension to the canonical version provided in
+ * zend-expressive-router v2.4 and up, and is deprecated in favor of that
+ * version starting in zend-expressive 2.2.
  *
- * When invoked, it will create an empty response with status code 200.
- *
- * You may optionally pass a response prototype to the constructor; when
- * present, that instance will be returned instead.
- *
- * The middleware is only invoked in these specific conditions:
- *
- * - a HEAD request
- * - with a `RouteResult` present
- * - where the `RouteResult` contains a `Route` instance
- * - and the `Route` instance defines implicit HEAD.
- *
- * In all other circumstances, it will return the result of the delegate.
- *
- * If the route instance supports GET requests, the middleware dispatches
- * the next layer, but alters the request passed to use the GET method;
- * it then provides an empty response body to the returned response.
+ * @deprecated since 2.2.0; to be removed in 3.0.0. Please use the version
+ *     provided in zend-expressive-router 2.4+, and use the factory from
+ *     that component to create an instance.
  */
-class ImplicitHeadMiddleware implements ServerMiddlewareInterface
+class ImplicitHeadMiddleware extends BaseImplicitHeadMiddleware
 {
-    const FORWARDED_HTTP_METHOD_ATTRIBUTE = 'forwarded_http_method';
-
-    /**
-     * @var null|ResponseInterface
-     */
-    private $response;
-
     /**
      * @param null|ResponseInterface $response Response prototype to return
      *     for implicit HEAD requests; if none provided, an empty zend-diactoros
@@ -57,55 +32,17 @@ class ImplicitHeadMiddleware implements ServerMiddlewareInterface
      */
     public function __construct(ResponseInterface $response = null)
     {
-        $this->response = $response;
-    }
+        trigger_error(sprintf(
+            '%s is deprecated starting with zend-expressive 2.2.0; please use the %s class'
+            . ' provided in zend-expressive-router 2.4.0 and later. That class has required'
+            . ' dependencies, so please also add Zend\Expressive\Router\ConfigProvider to'
+            . ' your config/config.php file as well.',
+            __CLASS__,
+            BaseImplicitHeadMiddleware::class
+        ), E_USER_DEPRECATED);
 
-    /**
-     * Handle an implicit HEAD request.
-     *
-     * If the route allows GET requests, dispatches as a GET request and
-     * resets the response body to be empty; otherwise, creates a new empty
-     * response.
-     *
-     * @param ServerRequestInterface $request
-     * @param DelegateInterface $delegate
-     * @return ResponseInterface
-     */
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
-    {
-        if ($request->getMethod() !== RequestMethod::METHOD_HEAD) {
-            return $delegate->process($request);
-        }
-
-        if (false === ($result = $request->getAttribute(RouteResult::class, false))) {
-            return $delegate->process($request);
-        }
-
-        $route = $result->getMatchedRoute();
-        if (! $route || ! $route->implicitHead()) {
-            return $delegate->process($request);
-        }
-
-        if (! $route->allowsMethod(RequestMethod::METHOD_GET)) {
-            return $this->getResponse();
-        }
-
-        $response = $delegate->process(
-            $request
-                ->withMethod(RequestMethod::METHOD_GET)
-                ->withAttribute(self::FORWARDED_HTTP_METHOD_ATTRIBUTE, RequestMethod::METHOD_HEAD)
-        );
-
-        return $response->withBody(new Stream('php://temp/', 'wb+'));
-    }
-
-    /**
-     * Return the response prototype to use for an implicit HEAD request.
-     *
-     * @return ResponseInterface
-     */
-    private function getResponse()
-    {
-        return $this->response ?: new Response();
+        parent::__construct($response ?: new Response(), function () {
+            return new Stream('php://temp/', 'wb+');
+        });
     }
 }